alert_system/requirements.txt
confluent-kafka==2.6.0
mysql-connector-python
